+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.10.0] - 2021-02-15
 ### Added
 - **limitPerSec** new argument to limit the number of metric packs sent per second. Default is 10 packs that is equals to 10x1000=10000 metrics per second or 600000 mpm.
+- New metric **PacksOverflewError** exported as **packs_overflew_error**.
 
 ## [0.9.0] - 2021-02-09
 ### Changed